move other plugins to compatibility module
