+++ bukkit/compatibility/build.gradle.kts
@@ -2,6 +2,8 @@ repositories {
     mavenCentral()
     maven("https://repo.papermc.io/repository/maven-public/")
     maven("https://r.irepo.space/maven/")
+    maven("https://repo.extendedclip.com/content/repositories/placeholderapi/") // papi
+    maven("https://maven.enginehub.org/repo/") // worldguard worldedit
 }
 
 dependencies {
@@ -10,6 +12,11 @@ dependencies {
     compileOnly("dev.folia:folia-api:${rootProject.properties["paper_version"]}-R0.1-SNAPSHOT")
     // NeigeItems
     compileOnly("pers.neige.neigeitems:NeigeItems:1.21.42")
+    // Placeholder
+    compileOnly("me.clip:placeholderapi:${rootProject.properties["placeholder_api_version"]}")
+    // WorldEdit
+    compileOnly("com.sk89q.worldedit:worldedit-core:7.2.19")
+    compileOnly("com.sk89q.worldedit:worldedit-bukkit:7.2.19")
 }
 
 java {
